GPT_LLM_Query_System/requirements.txt
fastapi
uvicorn
gunicorn
transformers
torch
pytesseract
pdf2image
pillow
nltk
spacy
pandas
llama-cpp-python
PyMuPDF
python-multipart
logging_loki
prometheus-client
prometheus-fastapi-instrumentator
